requirements.txt
djangorestframework
djongo
